Switch over to the new react event collection page
