+++ src/lib/urls.js
@@ -225,10 +225,6 @@ module.exports = {
     ),
   },
   events: {
-    // React routes (Work in progress) - to be released on completion
-    react: {
-      index: url('/events', '/react'),
-    },
     index: url('/events'),
     create: url('/events/create'),
     details: url('/events', '/:eventId/details'),
